Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 int main(){
     int r;
     scanf("%d",&r);
-    int a = 3.14*r*r;
+    int a = (int)(3.14 * r * r);
     printf("Area: %d",a);
     return 0;
     }
